Start typing transport.py based on transport.ts
This is just the beginning of an exploration into python3 typing for
the transport "interface". I found this PEP interesting:
https://www.python.org/dev/peps/pep-3119/#abcs-vs-interfaces
not sure if abc is really what we want but I'd rather not use
zope.interface (I think we should try to keep dependancies small).

I've added a "configure" method to the transport interface as a low cost
implementation of configuring dependancy injection.

TODO:
- determine if we use dataclasses versus attrs
- flesh out transport.py interface/types
- actually implement and test requests_transport.py