Serialize module
relies on caller to supply the top level model and wether it's a list. I
tried inferring this from the return type of the caller (which the
caller would still need to pass in). That would be DRYer but the typing
stuff got really hairy. Since we generate the code anyway I figure a
couple extra arguments to pass down the chain isn't too bad.