Rename LookerSDK to looker_sdk for pypi package
- more packaging prep: added looker-sample.ini, switched README
extension to .rst (since this is the default format pypi uses anyway)
- Serialization bug: instantiating a "Write*" object defaults all fields
to None and any that you don't set with appropriate values get sent
along in the json. For example, in an "update_user" call, specifying
only WriteUser(last_name="Doe") would erase all the other properties
(because the object has them set as None and they go with the json).