Python: simplify sending json null implementation
Upon Model instantiation all optional fields are assigned None when not
passed a specific value: standard behavior. When cattr serializes to
JSON, it ignores an fields with a value of None. In order to send a JSON
null value for a field it must be assigned the value of EXPLICIT_NULL