Python: stamp looker/api version for user-agent
For the codegen this is currently almost an exact copy of the typescript
generator with variable names snake_cased for python style. Might be
worth abstracting those variables out into config and moving the method
implementation into the base class.

For the python rtl - added User-Agent header like "PY-SDK 3.1.6.19". We
might want to also consider the same prefixing strategy we plan for
environment variable config here as well. (e.g. we'd be "LOOKER-PY-SDK
3.1.6.19")