fix for netrc creds overriding auth headers
As per https://3.python-requests.org/user/quickstart/#custom-headers, the requests library overrides auth headers with credentials specified in netrc if the auth parameter is not specified when calling requests.

The fix for now (until we support netrc files), is a NullAuth class that prevents this from happening, effectively ignoring the presence of netrc files on the system.