Fixing the legacy generator
The legacy generator version was also updated and it now complains about the `-DpackageName` option so it's less ideal.

The generator works, so I'm pushing the fix anyway. This is mainly to unblock someone wanting to use the legacy generator for `C#`

I also updated the versions on most of the node packages.