Python SDK: Add constructors for all input models
The logic for when to add a constructor to a model was flawed - "Write"
types are just copies of Models that have at least one read-only field.
Not every method gets a separate "Write" model.

This change also captures nested model types for input models to
generate constructors for those as well.

Another option would be to just generate a constructor for all models
but this way we at least still keep the code size down and once the IDE
support is there for attrs auto-generated constructors we can ditch them
all together.