Add transport and session implementations for the extension SDK
The SDK remains decoupled from both the extension and chatty packages. The extension SDK will provide an implementation of a new interface exported by this change. Note streams are not currently supported. Note the interface could be implemented by some other package so it begs the question as to whether extension is a good name. On the other hand, I can't see anything else wanting to use the new interface so I think the extension prefix is reasonable as a name.