Fix typescript test failures when tests run from root directory
Test failures were caused by tslint issues. When the tests were added, the tests were run from the typescript/looker directory and worked. This is because the tsconfig at root and in typescript/looker were different. This change makes an attempt at reconciling the differences. The format of the two tsconfig files are now similar so they can be compared side by side. Matching options are at the top, options with different values next, followed lastly by different options.

The options causing the test failures were:
    "noUnusedLocals": true,
    "noUnusedParameters": true,
These were not in the tsconfig in typescript/looker. Fixing not an option as this caused the typescript/looker build to fail. Also tried tslint disable but for some reason jest did not honour these. In the end removed those options in the root tsconfig to bring into line with typescript/looker tsconfig. yarn test from root no longer fail due to typescript issues.

Note that yarn lint fails from root (fine in typescript/looker). sdkConfig.ts has this line which fails with ERROR: 45:9  strict-type-predicates  Expression is always false.
 if (props.verify_ssl === undefined) props.verify_ssl  = true
I suspect this is because verify_ssl is a required property and as such should never be undefined. The code seems a little odd anyway but I would prefer to fix it in a separate PR.