Typescript multi-API SDK is working
There's now `Looker40SDK` and `Looker30SDK` and `sdk/3.1` and `sdk/4.0` paths for the API-specific SDK code

Still need to work on whether AgentTag is specified in the optimal place yet. I'm thinking it's not. There are some other, hopefully cleaner, ways of specifying it to the general-purpose request and stream functions.

However, this pattern is far enough along we can start making the multi-API SDK objects for Python, Kotlin, and Swift.