Python OAuth session support
Added OAuth session whose `_login` method calls `/api/token` with a
`refresh_token` grant_type. This assumes that you've set `self.token` with
an access token that has a `refresh_token`. We've provided a convenience
function to do this: `redeem_auth_code` which takes a temporary code
you've been given (see below) and calls `/api/token` with an
`authorization_code` grant_type.

In order to get the temporary code you must ask the authorization server
for it which will usually require a login, a redirect to authorize the
access, and then a final redirect to the location you supplied in the
initial request with a `code` query string parameter. For Looker
authorization we've provided another convenience function:
`create_auth_code_request_url` which will construct a URL with the
appropriate query string parameters to the Looker authorization server.

PS tests/integration/api40/test_methods.py tests are broken because the
spec changed. The failing tests are not related to this change. I'll
open a separate PR to fix them