Fixes to support embedding in other projects
This defines a project group and name which are necesssary so that the
kotlin sdk may be accessed as a project.  Until we have the project
published to a package distribution system, this allows us to use a
Gradle Source Dependency [1] to embed the project.

Also changes a few `compile` dependencies to their newer versions to
remove deprecation warnings from gradle.

[1] https://blog.gradle.org/introducing-source-dependencies