Route Param Validation (#344)
Example Google Access Token Server

A server that can request a Google access token using a service account. Looker credentials and a scope must be provided in order to get the access token. The access token can then be used to access google APIs that support updates.

Written for the hackathon application that stores data in a google sheet.