Refactor Kotlin RTL
The goals of this PR are twofold:
- Improve the handling of errors
- Improve the typing of responses

This mostly revoles around replacing SDKResponse with
SdkResult<TSuccess, TFailure>.  This allows us to track the type
information with the SdkResult and easily get out the expected type on
success or failure.

For now, all errors are expected to be of type com.looker.sdk.Error,
even though we do have some methods which return more precise errors.
If for some reason you need to escape the default typing, such as to get
a finer Error type, you can use `bodyAs<T>()` on both SuccessResponse
and FailureResponse.

For clients, this commit keeps the concept of the `ok()` method that
returns the success type or throws an error.  This moved from being a
method on the sdk itself to an extension function on the SdkResult,
so you can do `val foo: Foo = sdk.get_foo().ok()`.  If you want to do
more precise handling, SdkResult is a `sealed` class, so you can use a
`when` statement to check for the various type (SuccessResonse,
FailureResponse, Error) and get smart casting to access the appropriate
types and Response metatdata.