Bump kotlin version to 1.4.20
We want to explore kotlinx.serialization, but that requires Kotlin 1.4.
Bump the version now.

The additional change to the dependency declaration resolves some
warnings about having mismatched dependency jars, presumably from before
the upgrade.  This is likely not an issue on a clean build, but better
to get it straight for developers who are upgrading and existing
environment.