Upgrade kotlin version to 1.4.20 (#393)
* Upgrade gradle wrapper from 6.5.1 to 6.7.1

* build.gradle cleanup

Use the most current syntax for applying the kotlin plugin, using a
properties file to specify the version and the `plugins{ }` block
instead of the old `apply plugin` syntax.

* Bump kotlin version to 1.4.20

We want to explore kotlinx.serialization, but that requires Kotlin 1.4.
Bump the version now.

The additional change to the dependency declaration resolves some
warnings about having mismatched dependency jars, presumably from before
the upgrade.  This is likely not an issue on a clean build, but better
to get it straight for developers who are upgrading and existing
environment.

* Improve error logging in tests

Tests were failing with just short errors which weren't very helpful.
This changes the test reporter to use longer error messages for better
diagnostics of the failing tests.