Do not auto run CI on forked PRs
I misunderstood the security boundary for the pull_request_target event.
Apparently we should not be running code from forked PRs as it can gain
access to the github token which has access to our github secrets.

https://securitylab.github.com/research/github-actions-preventing-pwn-requests

I was made aware of this issue when we received this PR:
https://github.com/looker-open-source/sdk-codegen/pull/429/files

It attempted to exfiltrate git config and environment secrets from the
github runner. I reproduced the behavior sending the "loot" and
"loot-env" files to a request-bin I setup. After examining the contents
I'm confident that no secrets were exfiltrated. Regardless I will be
changing the GCP_* secrets as an extra precaution.