Eslint fixes (#463)
This fixes some but not all eslint issues. In some cases eslint disable has been used. Reasons any one of the following

1. lint rule does not make sense for code. example prefer-required, prefer to have class, prefer in document and import no extraneous dependencies (this one is weirs suspect eslint and lerna at cross purposes)
2. fix breaks test - example DocMarkdown.spec.ts
3. fix change to impactful and no really good test (api-explorer/scripts/utils.ts for example)

Note one fix removed a duplicate case statement (data/projects/reducer.ts). Probably not an issue but could be the wrong action or could be that the value should be cleared. Will wait for a bug to be reported.

Note CI now runs lint:es so lint errors need to be fixed now