Fix tests
Looker components now need to be transpiled for jest tests as they are now esm modules. The technique to do this was acquired from the main Looker repo. It uses babel.common.js which is also imported into webpack. No need for this build to import into webpack but following the same pattern.