chore: Looker components upgrade (#598)
* Looker components upgrade

Impacts the following
1. Icons changed - named strings no longer supported - need to provide an Icon component.
2. Tree control - issues with height, especially collapsing/expanding nodes.
3. List control - issues with height and spacing. In once case I had to replace List/ListItem with ui/li.
4. Data grid now requires a caption

* Fix tests

Looker components now need to be transpiled for jest tests as they are now esm modules. The technique to do this was acquired from the main Looker repo. It uses babel.common.js which is also imported into webpack. No need for this build to import into webpack but following the same pattern.

* Backout increase in node memory size. Not sure its helping

* remove package that does not have tests

* jest speed up hopefully

* Updated examples index

Co-authored-by: John Kaster <kaster@google.com>