chore: fix annoying warnings from APIX builds (#649)
Standalone and extension APIX developer and bundle builds were displaying import/export warning messages. This change fixes those warning messages.

The downside to this change is that we may be exporting functions and types that may not need to be exported. There should be another exercise to go through and clean up index files for lower level components such that only types and components that actually need to be exported are exported.