ci: centralize noop checks
The benefit of having them duplicated in each workflow
file was the ability to delay their execution with the
`needs: [publish-test-results]` syntax. However, there
was buggy behavior where the python-ci workflow would
get attributed the actual jobs from the tssdk-ci workflow
and its own noop jobs would somehow fail with the
opaque `No such file or directory` error

The downside of centralizing them is losing the ability to
specify that `needs` syntax. I believe that means they
will "pass" immediately and there will be a window of
several minutes until the "real" jobs report where the PR
will satisfy all the required checks and could be
mergeable. If that is the case, I may consider hacking it
to do `sleep 300 && exit 0` which would introduce an
artifical delay that may or may not help avoid the issue