feat: upgrade react-markdown (#681)
* feat: upgrade react-markdown

upgrade required for the following reasons
1. dependabot reported a security issue with a package (trim) the older version of react-markdown was using. Note that the trim packages is still referenced by eslint tooling. This will be dealt with in a separate PR.
2. dependency on path-browserify and process-browser can be removed.

The replacement was not straight forward as the mechanism for providing alternative rendering of components significantly changed. The Markdown component was refactored into its own component as react-markdown was being used in two places. It turns out that this was not necessary as the 2nd usage needs to use the raw react-markdown component at least for. The refactor was kept as it is possible the markdown component may be useful elsewhere. The original implementation had a dependency on specKey which has now been removed.

In addition the following changes have been done:
1. Links to looker documentation in an extension now supported.
2. List components render correctly.
3. Links that have markup from search now work.