feat: initial api explorer dev portal support (#748)
# Description 

Allows API explorer to run without a banner inside of the developer portal. Note that this will eventually be refactored to build a version of API explorer inside of the developer portal. For now API explorer is built in sdk codegen and the resultant javascript build is copied to the developer portal.

# Changes

1. Looks for LOOKER_DEV_PORTAL object to get router basename and to determine if to run headless.
2. Headless parameter added to API explorer. When running in headless mode, selectors are moved to the sidebar. Current header is not shown.
3. Selectors have been wrapped to allow simplified use in two places (header or sidebar)
4. Tests updated

# To dos

1. Specific tests for headless
2. Externalize sidebar toggle such that dev portal can open/close API explorer sidebar (diff closes sidebar so a page reload is required to redisplay sidebar)
3. Fix diff navigation - occasionally results in an error or does not work.