feat: apix sidebar toggle for dev portal (#751)
# Description

This change allows an entity other than APIX explorer to hide/show the APIX sidebar.

# Changes

The dev portal navigation bar is independent of the APIX nav bar. A previous change, allowed the APIX nav bar to be hidden when APIX is running in the dev portal. The issue with that change  was that the toggle to open/close the sidebar was no longer visible and the toggle in the dev portal had no way to communicate with the APIX explorer so that it could hide/show the side bar. This change adds a window message listener to APIXExplorer when it operates in headless mode. When the window message is received it toggles the sidebar visibility.