fix: Extension Api Explorer markdown link clicks
Clicking a link in markdown now works for extensions.

Fix is to rely on getAttribute method to get pathname and href values rather
than get pathname and href directly. Also, now tests if url begins with
the spec key as well as the pathname.

Possibly broken by markdown upgrade although why standalone was not
broken is puzzling