chore: move search state into redux (#868)
* move search state to redux
- refactor various interleaved side nav states into one
- update SideNav useEffect deps list to be exhaustive

* make useDebounce a generic so at to avoid unnecessary non-null assertions on result (debouncedPattern)

* move spec.api assignment into useEffect

prior to this const api = spec.api resulted in a different object on
every rerender and since api was part of the effect's dependency list it
was causing an infinite rerender

* remove SearchCriteriaSelector

This component is unused and references code that no longer exists. SearchCriteria are part of the settings store so reimplementing this is trivial.