fix: hackathon build
Failing because of typescript errors in extension-utils package tests.

Changes:

1. Added new script to explicitly build ALL extensions (uses a slightly
different build than module so this should catch errors that may evade
the module build).
2. Added packages/extension-utils to apix-tests (changes to
extension-utils triggers apix-ci so this is okay).
3. Added build-extensions to apix-ci. Should catch errors in that future
that this PR fixes.
4. Fixed extension-utils package.json. Copy pasta error where it was
trying to be an extension and its not.
5. Fixed type issues extension-utils tests. Most likely caused by
changes being made for the apix explorer refactor. Missed as ci not
runnning the tests. That barn door has not been closed.