chore: skip ci tests but return success if files not modified (#947)
Prior to this change the some required tests wouldn't run. For instance, if Python code was changed, the TypeScript CI tests would not run at all. If the TypeScript CI tests were required to allow the PR to be merged, the PR would be blocked.

Instead of firing the CI tests on only some pushes and pull requests fire them on all, but skip all processing.

Instead of using a list of paths to trigger the workflow, the workflow is triggered for all pushes and merges. The list of paths is checked and a flag set. Subsequent steps are only run if files within those paths are changed. Otherwise they are skipped.