fix: route from diff page
Setting of current spec has been hoisted to ApiExplorer. ApiExplorer monitors the current
route and sets the current spec if route is different from current spec. Added a check in
the rendering to make sure that the current spec is set correctly before it allows the
router component to render. Renders the loading component instead which I think is okay
but may need to be revisited.

ApiSpecSelector has been modifief to ONLY change the route.

This means that if a component needs to switch spec all it needs to do is modify the route
(diff component showing detail for example).