chore: move specs into redux and cleanup oauth related code (#976)
* replace runItSettings and configurator with OAuthConfigProvider

* move specs to redux and cleanup oauth related code
- Work around non-reentrant code
- Cleanup OAuthScene and more tweaks to initSpecs
- Better memory handling
- chore: clean up extension-utils package.json
- feat: move specs to redux

* fix: type and diff scene related issues

* fix: stop diff links from reloading

* fix: DiffScene back/forward browser button behaviour
DiffScene did not handle changes in route when back or forward button
pressed. This change has the react router driving the changes to state.
If diff changes, history is updated. History updates state. When back/forward
button pressed, history is update and again history updates state.

* fix: route from diff page
Setting of current spec has been hoisted to ApiExplorer. ApiExplorer monitors the current
route and sets the current spec if route is different from current spec. Added a check in
the rendering to make sure that the current spec is set correctly before it allows the
router component to render. Renders the loading component instead which I think is okay
but may need to be revisited.

ApiSpecSelector has been modified to ONLY change the route.

This means that if a component needs to switch spec all it needs to do is modify the route
(diff component showing detail for example).

* fix: initial load issue

Co-authored-by: Bryn Ryans <bnryans@google.com>
Co-authored-by: John Kaster <kaster@google.com>