chore: Release extension-playground (#1251)
extension-playground's package.json is set to:
```
    "@looker/extension-sdk": "^22.4.2",
    "@looker/extension-sdk-react": "^22.4.2",
    "@looker/sdk": "^22.4.2",
```
We need release-please to update it's package.json or else what happens is right now, on calling `yarn`, we download the pinned versions and they are added to our lock file since the latest calendar version is 23.0.0 which doesn't work with the caret `^` that pins to the 22 major version. This probably pulls in the old build issue. I'm hoping this fixes that. 

This change will also release extension-playground as 1.0.0.