fix: Make Go SDK set correct content-type header (#1283)
This change makes Go SDK set Content-Type header according to the body's type. Either text/plain if it is a string, or application/json if it is JSON formatted string. The oauth2 transport automatically sets a url encoded content type header for login call. This change will fix requests 500'ing on % being in the json formatted body.

Fixes issue #1075