Create NetStandard 2.0 library
Split the library in two parts: One runtime library which targets .NET Standard 2.0 and one test library that targets .NET 6 and consumes the runtime library. This way, Xunit and other test-related libraries are not required to be deployed when referencing Look#.

It is recommended to use the lowest/oldest .NET Standard version which under which your library can execute to maximize compatibility.

In addition, add support for setting custom headers on a request-by-request basis. This is required because Looker requires the User-Agent header to be present when acquiring an embed session.