feat: update Go SDK to accept auth session interface (#1488)
This change alters the Go SDK client struct to accept an interface
instead of the AuthSession. The interface allows for custom
implementations of the RTL package, but since the existing AuthSession
fulfills the new interface (its based on the Do method from the current
AuthSession after all) this change will be completely transparent to
those who are currently using the current AuthSession.