chore: Added packages.lock to csharp sdk
Generated a project locklife for the csharp sdk to lock down package dependencies.
This was done to remediate an internal vulnerability scan ticket and to allow
Google to manage our csharp supply chain vulnerabilities.

Currently afaict, we dont have any CI running for csharp given its community supported. So there's nothing in the repo that relies on the lock file.
The lockfile shouldn't need to be updated when generating a csharp sdk as we only generate api/model definitions which shouldn't depend on the csharp sdk's packages.