chore: Added packages.lock to csharp sdk (#1621)
Generated a project lock file for the csharp sdk to lock down package
dependencies.

This was done to remediate an internal vulnerability scan ticket and to
allow Google to scan/manage our csharp supply chain vulnerabilities.

Currently afaict, we dont have any CI running for csharp given its
community supported, so there's nothing in the repo that relies on the
lock file. The lockfile shouldn't need to be updated when generating a
csharp sdk as we only generate api/model definitions which shouldn't
depend on the csharp sdk's packages. Example generated PR:
https://github.com/looker-open-source/sdk-codegen/pull/1612/files